attendify_django/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-decouple==3.8
PyJWT==2.10.1
bcrypt==4.3.0
cloudinary==1.44.1
python-dotenv==1.1.1
dj-database-url==3.0.1
psycopg2-binary==2.9.10
django-environ==0.12.0
pymongo==4.10.1
djongo==1.3.6
